+++ Dockerfile.openpilot
@@ -81,6 +81,8 @@ COPY ./pyextra /tmp/openpilot/pyextra
 COPY ./panda /tmp/openpilot/panda
 COPY ./external /tmp/openpilot/external
 COPY ./tools /tmp/openpilot/tools
+COPY ./laika /tmp/openpilot/laika
+COPY ./laika_repo /tmp/openpilot/laika_repo
 
 COPY SConstruct /tmp/openpilot/SConstruct
