+++ selfdrive/locationd/kalman/.gitignore
@@ -1,5 +1,1 @@
-lane.cpp
-gnss.cpp
-loc*.cpp
-live.cpp
-pos_computer*.cpp
+generated/
