+++ selfdrive/locationd/kalman/templates/compute_pos.c
@@ -1,13 +1,15 @@
-#include <eigen3/Eigen/QR>    
-#include <eigen3/Eigen/Dense>    
-#include <iostream>    
+#include <Eigen/QR>
+#include <Eigen/Dense>
+#include <iostream>
+
 typedef Eigen::Matrix<double, KDIM*2, 3, Eigen::RowMajor> R3M;
 typedef Eigen::Matrix<double, KDIM*2, 1> R1M;
 typedef Eigen::Matrix<double, 3, 1> O1M;
 typedef Eigen::Matrix<double, 3, 3, Eigen::RowMajor> M3D;
 
+extern "C" {
 void gauss_newton(double *in_x, double *in_poses, double *in_img_positions) {
-  
+
   double res[KDIM*2] = {0};
   double jac[KDIM*6] = {0};
 
@@ -28,7 +30,7 @@ void gauss_newton(double *in_x, double *in_poses, double *in_img_positions) {
 
 void compute_pos(double *to_c, double *poses, double *img_positions, double *param, double *pos) {
     param[0] = img_positions[KDIM*2-2];
-    param[1] = img_positions[KDIM*2-1]; 
+    param[1] = img_positions[KDIM*2-1];
     param[2] = 0.1;
     gauss_newton(param, poses, img_positions);
 
@@ -48,4 +50,5 @@ void compute_pos(double *to_c, double *poses, double *img_positions, double *par
     O1M ecef_output(pos);
     ecef_output = rot*ecef_output + ecef_offset;
     memcpy(pos, ecef_output.data(), 3 * sizeof(double));
+}
 }
